tests: fix test logic for pcre-context
Included in the original bug #20957, but corrupted somehow in
transit as the required NUL characters are missing.

Add a simpler version of the test case that uses plain characters
and match the -z data and output to show the equivalence.

Note the output is still not correct as it is missing the expected
LF characters, but a full fix will have to wait until PCRE2.

Fixes Bug#51735.